ITD/backend/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-seed==0.2.2
djangorestframework==3.12.2
Faker==5.7.0
gunicorn==20.0.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
